fixed resubmit data bug
